Add spacing to header and main section
Set the flexbox container with proper spacing for the header and top
body elements. As of now, no flexbox on the column direction has been
added to <body> and it seems like it won't need it.

Spacing between edge and header elements set as 20px for top, and 300px
for the sides. The sides spacing should be followed for the rest of the
sections as to keep an uniform look.